refactor: flatten utils
