+++ src/alert-on-update.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import fetch from 'unfetch';
 import ms from 'ms';
-import { isMobile } from './utils';
+import { isMobile } from './is-mobile';
 import { useUpdatePoller } from './use-update-poller';
 import UpdateAlert from './update-alert';
