--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './classify-characters';
-export * from './is-mobile';
